+++ docs/CHANGELOG.md
@@ -9,6 +9,8 @@
   - `wait_time`, which defines how long to wait after activation before beginning to move
   - `travel_distance`, which defines how many clicks the lift will travel vertically
   - `speed`, which defines how many world units the lift will travel per frame
+- added properties to the `O_BIG_BOWL` object:
+  - `pour_time`, which defines how long to pour liquid from the bowl
 - changed `Game mode selection` to allow hiding the dialog even after having completed the game (Gameplay → General → Game mode selection)
 - improved handling of dead enemies used as switch triggers in custom levels by not altering their activation status and by adding detection for having been exploded (#5682)
 - fixed Lara having an empty left holster if the option to remember guns between levels is used and she finished with the Desert Eagle (#5697)